refactor: extract the frontmatter codec into its own component (PR3, task 4a3836a9)
Split the corpus file format — Markdown+YAML <-> KnowledgeDocument — out of
knowledge.py into a new pure module, lithos.frontmatter_codec, with decode/encode
as its whole interface. knowledge.py shrinks 2444 -> ~1720 lines and keeps the
store (CRUD, metadata cache, slug/path allocation, reconcile seam).

graph.py and provenance.py need the document *type*, not the *manager*. While the
type lived in knowledge.py, importing it dragged in KnowledgeManager — the edge
that wired Graph/Provenance/Search into a cycle with Knowledge. They now import a
dependency-free leaf, which breaks that cycle:

  module_cycles              2 -> 1   (knowledge<->graph<->search<->provenance gone)
  component_cycles           1 -> 0
  cross_component_edges     63 -> 68   (budget raised; see below)
  cross_module_private_refs 38 -> 34

The codec gets its OWN architecture component rather than folding into Knowledge.
No import differs between the two mappings — but filing it under Knowledge would
keep the map drawing Graph->Knowledge / Provenance->Knowledge edges that no longer
exist at the module level, asserting a cycle the extraction already broke. Its own
component tells the truth (component_cycles 1->0) at the cost of 5 more
cross-component edges: a component importing both the store and the format now
counts as two edges where it was one. The budget raise is deliberate and reviewed.

frontmatter_codec sits in the Core tier but is held to Foundation discipline by
import-linter, as lithos._merge and lithos.edge_store are; its generated component
page shows "Depends on: -".

Notable:
- decode(text, relative_path) takes text, not a path — the codec never touches a
  disk; the manager reads bytes and hands them over. The round-trip law is testable
  with no manager and no filesystem.
- The two near-duplicate decode blocks in knowledge.py (read() and
  _sync_from_disk_unlocked) collapse to one decode() call each; read()'s truncation
  still happens after decode, so an excerpt never changes which links a caller sees.
- to_markdown() (2 sites) -> encode(doc); no BC shim.
- _normalize_datetime -> public normalize_datetime (was reached across 4 module
  boundaries).
- generate_slug deleted — a pure alias for slugify with zero production callers.

New tests/test_codec_conformance.py (23 tests): round-trip law, unknown-key
preservation, H1 title inversion, path-derived LCMA defaults, timezone
normalisation, strict-write/lenient-read confidence asymmetry, purity.

make check (1592 passed) + make test-integration (454 passed) + make diagrams
(48 guardrails) + lint-imports (3 contracts kept).
