+++ src/lithos/tools/findings_stats.py
@@ -11,7 +11,7 @@ from fastmcp import FastMCP
 
 from lithos.envelopes import invalid_input_envelope
 from lithos.events import FINDING_POSTED, LithosEvent
-from lithos.knowledge import _normalize_datetime
+from lithos.frontmatter_codec import normalize_datetime
 from lithos.telemetry import get_current_span, tool_metrics
 from lithos.tools._seam import tool_span
 
@@ -98,7 +98,7 @@ def register(mcp: FastMCP, server: LithosServer) -> None:
         since_dt = None
         if since:
             try:
-                since_dt = _normalize_datetime(datetime.fromisoformat(since))
+                since_dt = normalize_datetime(datetime.fromisoformat(since))
             except ValueError:
                 return invalid_input_envelope(f"Invalid since datetime: {since}")
